GH-20161: [Website] Remove mentions of master branch from Apache Arrow website content (#324)
# Overview

As part of apache/arrow#31142 and in response to the [recent rename of
the `apache/arrow-site` repository default branch to
`main`](https://issues.apache.org/jira/browse/INFRA-24242), this pull
request removes mentions of the "master" branch from the Apache Arrow
website content.

# Qualification

To qualify these changes, I:

1. Performed a local build and preview of the website using `bundle exec
rake` and manually verified that all the links are still working.

# Future Directions

1. There are still some remaining tasks listed in apache/arrow#31142
related to the renaming of the default branch of the `apache/arrow` repo
(and related repositories) to `main`.

# Notes

**Note**: I left the mentions of "master" in the `_release` and `_posts`
directory unmodified. As described in
https://github.com/apache/arrow/issues/20161#issuecomment-1442307810,
it's unclear whether it makes sense to change these uses since these are
referring to a past state of the repository. If the community would
prefer that we also removes mentions of "master" from these directories,
I would be more than happy to do so.

Closes apache/arrow#20161.