[R] Fix version selector in docs for 11.0.0 (#352)
The JS version selector was broken in the docs for 11.0.0, and there are
changes which have already been merged for later versions which replace
this with a link in the page footer to a versions list. However, we need
to add this for the archived version of the docs too, so it's not broken
there.

I did a 'find and replace' for the footer to link to the new
`versions.html` page (which will exist once the 12.0.0 docs are
published) and copied the new versions of extra.js so we no longer run
that code.