Bump .ruby-version from 3.0.6 to 3.3.6 (#566)
This repo includes a `.ruby-version` file and the Gemfile refers to
this. The site can be built with other Ruby versions but:

1. I think It's best to keep whatever version we specify in
`.ruby-version` to something that's [in
support](https://endoflife.date/ruby). Ruby 3.0.6 is not.
2. Our GHA workflows use
[`ruby/setup-ruby`](https://github.com/ruby/setup-ruby) which uses the
Ruby version specified in `.ruby-version` which is another reason to
specify an in support version of Ruby

I chose Ruby 3.3.6 because it's the latest release as of writing.

I've tested this PR locally by:

- Bundling under Ruby 3.3.6
- Generating the site (`bundle exec generate`)

And all looks good.