Try fix Content Security Policy errors (take 9) (#603)
This should be the final one of these PRs. It removes the dynamic GitHub
star button which included the star count (because this violated the
[ASF CSP](https://infra.apache.org/csp.html) by trying to retrieve
content from the [GitHub API](https://api.github.com/)) and replaces it
with an entirely static self-hosted button (like the LinkedIn and X
buttons). This also removes the last of the attempts to experiment with
setting a customer CSP, and just uses the default ASF CSP.