Website: Fix issues with R docs site (#614)
The current website has a CSS issue that looks like a regression in
pkgdown we didn't catch. This PR re-generates the site using the latest
pkgdown which fixes the issue for me locally. I caught two other things
while reviewing changes:

- Malformed text in authors.html
- Missing docs page for `one`. I think the search index database might
not include this with these changes but I'm not overly worried about it.
It'll get fixed in the v20 release.