Update .htaccess file to allow kapa.ai bot to work (#619)
This PR updates the .htaccess file to allow the kapa.ai bot to work
correctly. The [Infra guidance](https://infra.apache.org/csp.html) notes
that external resources are disallowed for providers we **do not** have
a DPA with, but kapa.ai **does** [have one with the
ASF](https://privacy.apache.org/faq/committers.html) so this should be
all good.

---------

Co-authored-by: Bryce Mecum <petridish@gmail.com>