GH-695: [Website] Fix anchors on visual identity page (#696)
I think the recent markdown parser change made it so we can't do headers
like this anymore,

```
# Design of the logo {#design}
```

This fixes the [current
rendering](https://arrow.apache.org/visual_identity/) which has headers
that look like this:

> Design of the logo {#design}

and restores the internal document anchors. A better solution might be
to look into a Jekyll plugin that automatically enables anchors on
headers and ToCs.

Closes #695